+++ VoodooInput/VoodooInputMT1Simulator/VoodooInputMT1UserClient.hpp
@@ -51,7 +51,7 @@ public:
     IOReturn sGetReport(void *p1, void *p2, void *p3, void *p4, void *p5, void *p6);
     IOReturn sNoop(void *p1, void *p2, void *p3, void *p4, void *p5, void *p6);
     
-    void enqueueData(void *data, UInt32 size);
+    void enqueueData(void *data, size_t size);
 private:
     VoodooInputMT1Simulator *simulator {nullptr};
